whatpm/t/tree-test-void.dat

#data
<command>a
#errors
no DOCTYPE
#document
| <html>
|   <head>
|     <command>
|   <body>
|     "a"

#data
<eventsource>a
#errors
no DOCTYPE
#document
| <html>
|   <head>
|     <eventsource>
|   <body>
|     "a"

#data
<event-source>a
#errors
no DOCTYPE
unexpected EOF
#document
| <html>
|   <head>
|   <body>
|     <event-source>
|       "a"

#data
<head><command>a
#errors
no DOCTYPE
#document
| <html>
|   <head>
|     <command>
|   <body>
|     "a"

#data
<head><eventsource>a
#errors
no DOCTYPE
#document
| <html>
|   <head>
|     <eventsource>
|   <body>
|     "a"

#data
<head><event-source>a
#errors
no DOCTYPE
unexpected EOF
#document
| <html>
|   <head>
|   <body>
|     <event-source>
|       "a"

#data
</head><command>a
#errors
no DOCTYPE
#document
| <html>
|   <head>
|   <body>
|     <command>
|     "a"

#data
</head><eventsource>a
#errors
no DOCTYPE
#document
| <html>
|   <head>
|   <body>
|     <eventsource>
|     "a"

#data
</head><event-source>x
#errors
no DOCTYPE
unexpected EOF
#document
| <html>
|   <head>
|   <body>
|     <event-source>
|       "x"

#data
<body><command>a
#errors
no DOCTYPE
#document
| <html>
|   <head>
|   <body>
|     <command>
|     "a"

#data
<body><eventsource>x
#errors
no DOCTYPE
#document
| <html>
|   <head>
|   <body>
|     <eventsource>
|     "x"

#data
<body><event-source>x
#errors
no DOCTYPE
unexpected EOF
#document
| <html>
|   <head>
|   <body>
|     <event-source>
|       "x"

#data
<head><noscript><command>x
#errors
no DOCTYPE
no </noscript>
#document
| <html>
|   <head>
|     <noscript>
|   <body>
|     <command>
|     "x"

#data
<head><noscript><eventsource>x
#errors
no DOCTYPE
no </noscript>
#document
| <html>
|   <head>
|     <noscript>
|   <body>
|     <eventsource>
|     "x"

#data
<head><noscript><event-source>x
#errors
no DOCTYPE
no </noscript>
unexpected EOF
#document
| <html>
|   <head>
|     <noscript>
|   <body>
|     <event-source>
|       "x"

#data
<b><i></b><area>
#errors
no DOCTYPE
no </i> before </b>
unexpected EOF
#document
| <html>
|   <head>
|   <body>
|     <b>
|       <i>
|     <i>
|       <area>

#data
<b><i></b><basefont>
#errors
no DOCTYPE
no </i> before </b>
unexpected EOF
#document
| <html>
|   <head>
|   <body>
|     <b>
|       <i>
|     <i>
|       <basefont>

#data
<b><i></b><bgsound>
#errors
no DOCTYPE
no </i> before </b>
unexpected EOF
#document
| <html>
|   <head>
|   <body>
|     <b>
|       <i>
|     <i>
|       <bgsound>

#data
<b><i></b><br>
#errors
no DOCTYPE
no </i> before </b>
unexpected EOF
#document
| <html>
|   <head>
|   <body>
|     <b>
|       <i>
|     <i>
|       <br>

#data
<b><i></b></br>
#errors
no DOCTYPE
no </i> before </b>
</br>
unexpected EOF
#document
| <html>
|   <head>
|   <body>
|     <b>
|       <i>
|     <i>
|       <br>

#data
<b><i></b><embed>
#errors
no DOCTYPE
no </i> before </b>
unexpected EOF
#document
| <html>
|   <head>
|   <body>
|     <b>
|       <i>
|     <i>
|       <embed>

#data
<b><i></b><img>
#errors
no DOCTYPE
no </i> before </b>
unexpected EOF
#document
| <html>
|   <head>
|   <body>
|     <b>
|       <i>
|     <i>
|       <img>

#data
<b><i></b><spacer>
#errors
no DOCTYPE
no </i> before </b>
unexpected EOF
#document
| <html>
|   <head>
|   <body>
|     <b>
|       <i>
|     <i>
|       <spacer>

#data
<b><i></b><wbr>
#errors
no DOCTYPE
no </i> before </b>
unexpected EOF
#document
| <html>
|   <head>
|   <body>
|     <b>
|       <i>
|     <i>
|       <wbr>

#data
<b><i></b><param>
#errors
no DOCTYPE
no </i> before </b>
#document
| <html>
|   <head>
|   <body>
|     <b>
|       <i>
|     <param>

#data
<b><i></b><source>
#errors
no DOCTYPE
no </i> before </b>
#document
| <html>
|   <head>
|   <body>
|     <b>
|       <i>
|     <source>

#data
<command/>
#errors
no DOCTYPE
#document
| <html>
|   <head>
|     <command>
|   <body>

#data
<eventsource/>
#errors
no DOCTYPE
#document
| <html>
|   <head>
|     <eventsource>
|   <body>

#data
<event-source/>
#errors
no DOCTYPE
/>
unexpected EOF
#document
| <html>
|   <head>
|   <body>
|     <event-source>

#data
<param/>
#errors
no DOCTYPE
#document
| <html>
|   <head>
|   <body>
|     <param>

#data
<source/>
#errors
no DOCTYPE
#document
| <html>
|   <head>
|   <body>
|     <source>
